requirements.txt
dj-database-url==0.5.0
Django==2.1.1
django-extensions==2.1.2
django-heroku==0.3.1
djangorestframework==3.8.2
gunicorn==19.9.0
ipython==6.5.0
ipython-genutils==0.2.0
psycopg2==2.7.5
pytz==2018.5
six==1.11.0
whitenoise==4.0
